restructuring of files and folders on the DDD design pattern
