+++ package.json
@@ -18,6 +18,7 @@
     "eslint-plugin-prettier": "^3.1.4",
     "prettier": "^2.0.5",
     "ts-node-dev": "^1.0.0-pre.51",
+    "tsconfig-paths": "^3.9.0",
     "typescript": "^3.9.6"
   },
   "dependencies": {
@@ -33,7 +34,7 @@
   },
   "scripts": {
     "build": "tsc",
-    "dev:server": "ts-node-dev --inspect --transpile-only  --ignore-watch node_modules --no-notify src/server/index.ts",
-    "typeorm": "ts-node-dev ./node_modules/typeorm/cli.js"
+    "dev:server": "ts-node-dev -r tsconfig-paths/register --inspect --transpile-only  --ignore-watch node_modules --no-notify src/shared/infra/http/server/index.ts",
+    "typeorm": "ts-node-dev -r tsconfig-paths-register  ./node_modules/typeorm/cli.js"
   }
 }
