+++ src/modules/users/http/routes/sessions.route.ts
@@ -1,5 +1,5 @@
 import { Router, Request, Response } from 'express';
-import AuthenticateUserService from '../services/AuthenticateUserService';
+import AuthenticateUserService from '@modules/users/services/AuthenticateUserService';
 
 const sessionRouter = Router();
