+++ src/modules/users/services/CreateUserService.ts
@@ -1,7 +1,7 @@
 import { getRepository } from 'typeorm';
 import { hash } from 'bcryptjs';
-import User from '../database/models/User.model';
-import AppError from '../errors/AppError';
+import User from '@modules/users/infra/typeorm/entities/User.model';
+import AppError from '@shared/errors/AppError';
 
 interface Request {
   name: string;
